Moved handling the EhrExtractRequest to new class
